requirements.txt
zeroconf
protobuf
ping3
cryptography
pyOpenSSL